[CLN] Refactor grids into its own modules
