+++ gempy/core/data/grid_modules/__init__.py
@@ -1,2 +1,4 @@
-from .grid_types import Sections, RegularGrid, CustomGrid
+from .regular_grid import RegularGrid
+from .custom_grid import CustomGrid
+from .sections_grid import Sections
 from .topography import Topography
